moved fit_pade to wicmth
